+++ tests/fixtures/rules/MLC104/helper.py
@@ -7,3 +7,14 @@ class HelperDemo(Scene):
 
     def construct(self):
         self.show(Square())
+
+
+class TwoSiteDemo(Scene):
+    def flash(self, mob):
+        self.play(mob.animate.shift(RIGHT), run_time=0)
+
+    def construct(self):
+        a = Square()
+        b = Circle()
+        self.flash(a)
+        self.flash(b)